Move metadata code to hosts package and do some cleanup
