+++ internal/hosts/metadata_kvstore_test.go
@@ -1,4 +1,4 @@
-package environments
+package hosts
 
 import (
 	"fmt"
